move transaction-executor to core
